pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "d4-market-bot"
version = "1.0.0"
description = "Automated trading bot for diablo.trade — multi-account, AI-powered, with Discord integration and a web dashboard."
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "loguru>=0.7.3",
    "aiosqlite>=0.22.1",
    "aiohttp>=3.13.3",
    "python-dotenv>=1.2.2",
    "google-genai>=1.66.0",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "discord.py>=2.7.1",
    "Levenshtein>=0.27.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
    "mypy>=1.19.1",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
